Use the new edit log system to add and remove key frame layers
Really need a way to do this via the layers themselves; thinking of passing in an edit log to the layer so we can supply edit commands there. Think we only need the ‘mutable’ side of the edit log there, which will simplify things a bit.

+++ animation/src/traits/edit/animation_edit.rs
@@ -11,7 +11,7 @@ pub enum AnimationEdit {
     DefineBrush(u32, BrushDefinition),
 
     /// Edit to an existing layer
-    Layer(u32, LayerEdit),
+    Layer(u64, LayerEdit),
 
     /// Sets the canvas size for this animation
     SetSize(f64, f64),
